feat(actionpack): port ActionDispatch::Routing::Endpoint
Extract the Endpoint base class into its own file mirroring Rails'
actionpack/lib/action_dispatch/routing/endpoint.rb layout. The class
was previously inlined in redirection.ts; api:compare now finds it at
routing/endpoint.ts to match the Rails source path.
